+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
-## v0.3.13
+## v0.4.0
 
-- Add Link as Button
+- Refactors sizes on Spacing component
+- Adds responsiveColumnReverse prop
+- Removes "both" direction
+- Sets "horizontal" as default direction
+- Adds textAlign prop on Spacing
+- Changes direction prop for vertical boolean prop